Changes easier compilation + minor extras:  * Find_packag'ing mujoco  * Change directories mujoco -> mujoco_ros (+ includes)  * gitignore top-level cmake + system-specific files
