+++ .gitignore
@@ -1,1 +1,9 @@
 *data/
+
+# Ignore OS/IDE specific-files
+.vscode/
+**/.DS_Store
+**/._.DS_Store
+
+# Ignore the Top-Level CMakeLists.txt
+/CMakeLists.txt
