--- CMakeLists.txt
@@ -1,1 +1,0 @@
-/home/aravind/mambaforge/envs/ros_env/share/catkin/cmake/toplevel.cmake
