+++ mujoco_ros/models/buddy.xml
