+++ mujoco_ros/models/meshes/mushr_wheel.stl
