+++ mujoco_ros/models/meshes/mushr_ydlidar.stl
