+++ mujoco_ros/models/mushr.xml
