+++ mujoco_ros/msg/MushrControl.msg
