+++ mujoco_ros/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>mujoco</name>
+  <name>mujoco_ros</name>
   <version>0.0.0</version>
   <description>The mujoco package</description>
